refactor(message): group playback modules into a playback/ subfolder
The three non-visual modules that turn a message into speech/audio
(use-message-playback, plan-playback, spoken-part-tracker) form a distinct
subsystem; nest them under message/playback/ so the folder root holds just
the message model and bar UI.
