+++ src/features/board/activation/button-activation.ts
@@ -1,7 +1,7 @@
 import { playAudio } from "@shared/audio/play-audio";
 import { speak } from "@shared/speech/speech-store";
+import type { UseMessagePlaybackReturn } from "../message/playback/use-message-playback";
 import type { UseMessageReturn } from "../message/use-message";
-import type { UseMessagePlaybackReturn } from "../message/use-message-playback";
 import type { UseBoardNavigationReturn } from "../navigation/use-board-navigation";
 import type { BoardAction, BoardButton } from "../types";
 import { resolveButtonIntent } from "./intent-resolver";
